솔루션 변경

+++ DX11_Wizard_of_Legend/ENGINE_SOURCE/jsConstantBuffer.cpp
@@ -7,14 +7,17 @@ namespace js::graphics
 	ConstantBuffer::ConstantBuffer(eCBType type)
 		: mType(type)
 	{
+
 	}
+
 	ConstantBuffer::~ConstantBuffer()
 	{
 	}
 
 	bool ConstantBuffer::Create(size_t size)
 	{
-		desc.ByteWidth = (UINT)size;
+		// 상수 버퍼
+		desc.ByteWidth = size;
 		desc.BindFlags = D3D11_BIND_FLAG::D3D11_BIND_CONSTANT_BUFFER;
 		desc.Usage = D3D11_USAGE::D3D11_USAGE_DYNAMIC;
 		desc.CPUAccessFlags = D3D11_CPU_ACCESS_WRITE;
